backend/requirements.txt
beanie
black
fastapi[standard]
firebase-admin
pdf2image
pydantic-extra-types
pydantic-settings
sendgrid
semanticscholar
boto3